fix(ui): improve `AssetGroup` and `Pill` component
