+++ packages/ui/src/AssetIcon/delegation-token.tsx
