+++ packages/ui/src/AssetIcon/group.tsx
@@ -1,13 +1,14 @@
 import { Metadata } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
 import { ElementType } from 'react';
 import cn from 'clsx';
-import { Size, AssetIcon } from './index';
+import { Size, AssetIcon, sizeMap } from './single';
 
 export interface AssetGroupProps {
   assets?: (Metadata | undefined)[];
   as?: ElementType;
   size?: Size;
   variant?: 'overlay' | 'split';
+  hideBadge?: boolean;
 }
 
 const OVERLAY_SIZE_MAP: Record<Size, string> = {
@@ -46,18 +47,19 @@ export const AssetGroup = ({
   as: Container = 'div',
   size = 'md',
   variant = 'overlay',
+  hideBadge,
 }: AssetGroupProps) => {
   if (variant === 'split') {
     return (
-      <Container className={cn('relative flex items-center gap-[1px]')}>
+      <Container className={cn('relative flex items-center gap-[1px]', sizeMap[size])}>
         {assets?.[0] && (
           <div className={cn(SPLIT_SIZE_MAP[size], LEFT_BADGE_SIZE_MAP[size], RIGHT_CLIP_PATH)}>
-            <AssetIcon metadata={assets[0]} size={size} />
+            <AssetIcon hideBadge={hideBadge} metadata={assets[0]} size={size} />
           </div>
         )}
         {assets?.[1] && (
           <div className={cn(SPLIT_SIZE_MAP[size], MARGIN_SIZE_MAP[size], LEFT_CLIP_PATH)}>
-            <AssetIcon metadata={assets[1]} size={size} />
+            <AssetIcon hideBadge={hideBadge} metadata={assets[1]} size={size} />
           </div>
         )}
       </Container>
@@ -67,7 +69,13 @@ export const AssetGroup = ({
   return (
     <Container className={cn('relative flex items-center', OVERLAY_SIZE_MAP[size])}>
       {assets?.map((asset, index) => (
-        <AssetIcon metadata={asset} key={index} size={size} zIndex={assets.length - index} />
+        <AssetIcon
+          hideBadge={hideBadge}
+          metadata={asset}
+          key={index}
+          size={size}
+          zIndex={assets.length - index}
+        />
       ))}
     </Container>
   );
