+++ packages/ui/src/AssetIcon/unbonding-token.tsx
