+++ packages/ui/src/Pill/index.tsx
@@ -72,7 +72,9 @@ export const Pill = ({ children, priority = 'primary', context = 'default' }: Pi
         getFont(context, density),
         getColor(priority, context),
         getBackgroundColor(priority, context),
-        'box-border inline-block max-w-full w-max rounded-full',
+        'box-border max-w-full w-max rounded-full',
+        'inline-flex items-center',
+        density !== 'slim' && 'gap-1',
         priority === 'secondary'
           ? 'border-2 border-dashed border-other-tonalStroke'
           : 'border-none',
